+++ philo_m/srcs/utils.c
@@ -6,13 +6,13 @@
 /*   By: athiebau <athiebau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/07 19:21:46 by athiebau          #+#    #+#             */
-/*   Updated: 2024/03/25 23:36:16 by athiebau         ###   ########.fr       */
+/*   Updated: 2024/03/27 02:07:42 by athiebau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/philosophers.h"
 
-size_t	get_time(void)
+long long	get_time(void)
 {
 	struct timeval	tv;
